Remove obsolete homepage reel helpers
